SAMSON-693 - Fix the compilation so that "-Wno-unused-but-set-variable" is only set on Linux
I have "simplified" the setting of the flags for compilation. For the overall
tree and the modules there are a base line set of flags to be paased to the
compiler. Then for each compilation mode; debug, release or coverage,
additional flags are appended to satisify each case. Then for Linux the flag
"-Wno-unused-but-set-variable" is applied using a test of CMAKE_SYSTEM_NAME to
determine if we are on Linux or not. Using "IF (NOT APPLE)" would apply to any
platform that is not OS X. Whilst this is not a problem right now it could be
in future.

You may need to delete BUILD_RELEASE / BUILD_DEBUG to get a working build.



git-svn-id: https://barricada.hi.inet/repositorio/samson/trunk@2590 4143565c-f3ec-42ea-b729-f8ce0cf5cbc3