First drop of UI tests
This payload contains 2 projects: TestUtils contains classes that
will be reused throught different tests. It contains a class that
abstracts the frontend through Selenium and a JobLauncher interface
which abstracts creation of Hadoop jobs at different levels (the
only implementation of this interface in this payload works at the
frontend level through Selenium).

The WebUI-tests project actually contains the tests.