Share root@andromeda01 SSH key
In order to make it easier to login as root among different machines
of Andromeda, default replacement files are set in a way that the same
SSH key is used for all machines in the cluster. Since this very key
was already present in authorized_keys file, it is possible for
any root user to login as root in any other machine of Andromeda cluster.