Move OpenVZ Puppet as a submodule of Cosmos
The OpenVZ modules of Puppet we were using are some additions to
a Libvirt Puppet module obtained from the Internet. Such module
is distributed under GPL license so we cannot include it in our
codebase (issue #1006). In addition, there is a lot of unused code
there that may be removed.

This change just moves the Puppet scripts we are already using as
a new submodule OpenVZ under Cosmos. These scripts were additions
authored by us, so there is no stolen code at all. The rest of
Puppet files are just removed.