Fix deployment to use the proper Ambari RPM repo
So far the Puppet scripts were using the same RPM repository for
fetching the Ambari packages. That repository was populated with
nightly builds of cosmos-ambari, what is appropriate for development
environments (vagrant and andromeda) but not for staging or production
(orion).

This commit fixes #985 by making repository URL a variable instead
of configuration constant, which is configured based on the target
environment.