Integrate OpenVZ image configuration into Puppet
So far, OpenVZ images was pre-configured in order to set the network
configuration, Yum repositories, and some other stuff. Such a process
was made by unpackaging the original VZ image, replacing the config
files, and packaging again. Two Bash scripts were available for that:
one to copy the common files and another one to copy the custom config
for each server. The image was unpackaged and packaged again twice, one
per script, with the consequent time both steps require.

This change moves the responsibility of configuring the images to Puppet
rather from the Bash files.