Execute domain provisioning in sequence
There was an attempt (see commit 3bf35b6) to refactor the libvirt domain
provisioning in order to execute different requests sequentially (not
concurrently). This was required in order to avoid two domains to be
created at the same time since that caused the MAC address generation to
produce the same values (see issue #1024).

Unfortunatelly, 3bf35b6 didn't resolve the problem. It guaranties the
sequential execution for the machines being provisioned in the same request,
but not in different requests being attended concurrently. That's the case
of `persistentHdfs` cluster, that requires machines of two different
profiles, which are processed in two concurrent and independen requests.

This change reverts the sequential execution of domain provisioning in
`LibVirtInfrastructureProvider` and defines a specific `ExecutionContext`
in `JnaLibVirtServer` to be used in domain creation. Such execution context
uses a thread pool with only one thread, so there is no concurrency at all
for domain provisioning actions.