Adapt line length to terminal in cosmos-cli
The `list` command of `cosmos-cli` had a fixed value for maximum
allowed line length. This caused some inconveniences as described
in issue #1103.

This change adds a new module `cosmos.common.terminal` with a
`get_terminal_size()` function, which returns the dimensions of the
terminal in a portable way (tested in OSX and Windows). That function
is now used to determine the maximum allowed length for the lines
while listing clusters.