Fix random failures in TcpServerTest
As reported in issue #1290, there are random failures in `TcpServerTest`
caused by `BindException` errors.

This change fixes the issue by:

* Selecting an available TCP port rather than a random one. This prevents
binding errors caused by concurrent executions of the test suite.
* Using `SO_REUSEADDR` flag for server connections. The method used to
select an available port is to bind a new socket to a port selected by the OS.
After that the port is recorded and the socket is closed. This port will be used
as an available one, but it may not be fully closed when the actual socket server
is created. To avoid such a problem, the address is reused.