Disable user when deregistered
User unregistration was not causing the user to be disabled on every
cluster he have SSH or HDFS permissions, as reported in issue #1509.

This change adds a new convenience function `disableUserFromAll` to
`ServiceManager` that disables the user from all running clusters. This
function is invoked from `UserUnregistrationWizard` when the user is
unregistered.