Eventually run TcpServerTest due to BindException
As reported in #1290, there are still some `BindException` errors
while executing `TcpServerTest`. There is still a remote possibility of
the random port calculated to be unavailable before the `acceptClient()`
function is invoked on the mock object. This is likely the reason of
the test to have failed again.

This change uses a `eventually` block to protect the test against these
random `BindException`s.