Fix again BindException error in TcpServerTest
The rest failed again. Makes sense, since the port is not regenerated
from one iteration of `eventually` to the next one. If it is occupied
the first time, it will likely be the next one.

This change fixes it by moving the eventually block to the whole test
rather than the `withServerMock` block.