Apply Special Case Pattern to Path
The initial design of `Path` class was not very clear. The case of a
root path that have no parent and no element name was mixed with the
case of a subpath that always have a parent and a element name.

This change applies the Special Case Pattern to `Path`, creating the
more specific entities `RootPath` and `SubPath`.