Add a `ServiceConfig` class to process HDFS service config
The basic operation of Infinity Server is to intercept the HTTP calls
to namenode, authenticate and authorize the requests and redirect the
invocation to the appropriate proxy of data node. In order to do all
this, it have to modify the endpoints returned by the name node and
replace them by the corresponding proxies.

This change provides a `ServiceConfig` class that wraps the HDFS
service configuration obtained from the `application.conf` file. Such
config may be used to determine how the endpoints have to be replaced
in the outgoing responses.