Make `RequestProcessorTest` fail on bind error
Some test cases of this class have failed due to timeouts while
expecting an actor message to be received. It is likely that the
reason is the HTTP server is not up and ready to receive connections.

This change checks the response of `IO(Http)` in order to determine
whether the bind operation succeeds. If not, the test would fail with
the corresponding assertion exception and we would know that’s the
actual reason of the message timeout.