Introduce cake pattern on infinity server
Define a number of components to be assembled into the infinity server:
InfinityAppComponent, AuthenticationComponent, AuthorizationComponent,
RequestProcessorComponent and ConfigComponent.

The new components are not functional, this change introduces
skeletons to implement them.

Some of the components produce Akka properties instead of actual
instances to allow actors to form a supervision hierarchy.

Also, rename requestTimeout to request-timeout setting for consistency
and create more packages to contain the components.