Move authentication and authorization spawn to request processor
So far the authentication and authorization actors were injected to the
request processor with the intention of wrapping them with pool routes.
After some analysis, this approach was discarded in favor of a single
pair of authentication/authorization actors for each request.

This change moves the authentication/authorization spawn to the request
processor with the corresponding changes in its testing.