Make `CosmosApiAuthentication` terminate when done
`RequestProcessor` now spawns an authentication actor that is used
to authenticate the request under processing. After authentication
is done, no more work is required do be done so no more messages are
expected.

This change makes `CosmosApiAuthentication` to terminate after the
authentication request has been processed. It means the actor is not
alive anymore from that point on, removing the unexpected behavoir
that would cause if it is not terminated.