Make RequestProcessor rely on action for dispatching
Enrich Action trait to have a run method able to service the request
and make RequestProcessor to invoke it after credentials have been
validated.  Also add the skeleton for the GetMetadata action to try
the design.

Some obsolete configurations relative to WebHdfs were removed as they
were an obstacle for this change.