Refactor error coding and rendering
The current error coding and rendering requires the exception types
to be aware of the rendering mechanism, and the error codes to be
factories of exceptions. That’s great for exceptions declared in the
`finatra` package, but it doesn’t fit well with other exceptions.

This change refactors some classes, moving them around, with the purpose
of having a error coding and rendering mechanism that fits for
exceptions declared in other packages. In addition. some classes that
were declared in inappropriate packages or had redundant names have
been moved and/or renamed.