[FIX] Fixed hidden transactionality bug.
Test "The wizard must reconfigure persistent HDFS cluster with current and deleted users" would fail only when executed individually.

The failure was due to to having a future-executor thread attempting to access a change introduced within the same transaction (atomic block) from the main thread.

Fortunately, this wasn't necessary because we already cater for the intermediate user Creating state. The transactions were split and the bug was resolved.