Partially implement GetMetadata action
This change provides a partial implementation for `GetMetadata` action.
It only works for files, but it establishes the basics to fully
implement it and implement other actions.

The Infinity configuration and a new URL mapper entity are made
available for actions to use it. This context should be enough to
implement any other action.