Improve package naming for infinity-common
There is a conceptual issue in `infinity-common` module. The types we
use to represent some basic entities of the filesystem, as file and
directory metadata, are considered as messages. That’s pretty weird
when we are implementing filesystem logic, returning types that are
conceptually considered messages of the communication protocol.

This change provides the following renames to mitigate these conceptual
issues.

* `message.json` is renamed as `json`
* `Path`, `PathMetadata` and `PathType` are moved into new package `fs`
* `messages.Action` is renamed to `messages.Request`. This avoids
a name conflict with the `Action` hierarchy in `infinity-server` module.
* `UserProfile` is moved to `permissions`