Provide an abstraction for HDFS name node capabilities
A new trait `NameNode` is provided in order to encapsulate the
capabilities of a name node server. The idea is to have a single
injectable object that makes possible to separate the actions from the
HDFS classes by consuming the name node operations as simple building
blocks.

This abstraction layer does not replace the `Action` hierarchy. This is
still useful to provide a command design pattern and to separate the
complex interactions that carry out some actions (e.g., remove a file
comprises checking permissions, remove the file and then return the
metadata of the new file location).

This change doesn’t include yet the refactor of any `Action` in order
to use `NameNode` trait. This will be implemented after this commit.