Introduce request-response logging in Metadata Server
The purpose of this change is to log a debug or warning line for each
successful or failed request, respectively. This line provides some
info on the original request and the response that was generated, very
useful for debugging errors.

This change requires some changes in the signatures of `Responder`,
`ExceptionRenderer` and `ActionResultHttpRenderer` in order to pass
the original request to the place where the response is generated. This
is the way that requests and responses are correlated.